Whitespace cleanup

+++ src/util/string2int.cpp
@@ -18,7 +18,7 @@ inline T str2number(const char *str, int base, bool safe)
 {
   int errno_bak=errno;
   errno=0;
-  char * endptr;
+  char *endptr;
 // _strtoi64 is available in Visual Studio, but not yet in MINGW
 #ifdef _MSC_VER
   const __int64 val=_strtoi64(str, &endptr, base);
@@ -28,20 +28,22 @@ inline T str2number(const char *str, int base, bool safe)
 
   if(safe)
   {
-    assert(0 == errno);
+    assert(0==errno);
     errno=errno_bak;
     assert(endptr!=str);
     if(std::numeric_limits<T>::min()==0)
     {
       // unsigned
-      assert(val >= 0);
-      assert((T)val <= std::numeric_limits<T>::max());
+      assert(val>=0);
+      assert(
+        (unsigned long long)(T)val<=
+        (unsigned long long)std::numeric_limits<T>::max());
     }
     else
     {
       // signed
-      assert(val <= (long long)std::numeric_limits<T>::max());
-      assert(val >= (long long)std::numeric_limits<T>::min());
+      assert(val<=(long long)std::numeric_limits<T>::max());
+      assert(val>=(long long)std::numeric_limits<T>::min());
     }
   }
